+++ packages/svelte-maplibre-storymap/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @undp-data/svelte-maplibre-storymap
 
+## 0.3.15
+
+### Patch Changes
+
+- a21106c: fix: removed margin-bottom from last slide of storymap. enable pointer-events for slide card.
+
+## 0.3.14
+
+### Patch Changes
+
+- bf3b33b: fix: adjusted storymap header style of margin and line-height
+- 127029d: fix: apply position:sticky for storymap's map part and show footer after map.
+
 ## 0.3.13
 
 ### Patch Changes